+++ nrf_802154/serialization/src/nrf_802154_spinel_app.c
@@ -572,6 +572,36 @@ bail:
     return cancel_result;
 }
 
+bool nrf_802154_receive_at_scheduled_cancel(uint32_t id)
+{
+    nrf_802154_ser_err_t res;
+    bool                 cancel_result = false;
+
+    SERIALIZATION_ERROR_INIT(error);
+
+    NRF_802154_SPINEL_LOG_BANNER_CALLING();
+
+    nrf_802154_spinel_response_notifier_lock_before_request(
+        SPINEL_PROP_VENDOR_NORDIC_NRF_802154_RECEIVE_AT_SCHEDULED_CANCEL);
+
+    res = nrf_802154_spinel_send_cmd_prop_value_set(
+        SPINEL_PROP_VENDOR_NORDIC_NRF_802154_RECEIVE_AT_SCHEDULED_CANCEL,
+        SPINEL_DATATYPE_NRF_802154_RECEIVE_AT_SCHEDULED_CANCEL,
+        id);
+
+    SERIALIZATION_ERROR_CHECK(res, error, bail);
+
+    res = net_generic_bool_response_await(CONFIG_NRF_802154_SER_DEFAULT_RESPONSE_TIMEOUT,
+                                          &cancel_result);
+
+    SERIALIZATION_ERROR_CHECK(res, error, bail);
+
+bail:
+    SERIALIZATION_ERROR_RAISE_IF_FAILED(error);
+
+    return cancel_result;
+}
+
 void nrf_802154_pan_id_set(const uint8_t * p_pan_id)
 {
     nrf_802154_ser_err_t res;